update expressjs-postgres example to typescript
