+++ examples/expressjs-postgres/package.json
@@ -1,17 +1,26 @@
 {
-  "name": "with-express",
+  "name": "expressjs-postgres",
   "private": true,
   "version": "1.0.0",
   "main": "index.js",
   "author": "Jake Runzer",
   "license": "MIT",
   "scripts": {
-    "start": "node src/index.js",
-    "build": "echo 'No build step'"
+    "dev": "nodemon src/index.ts",
+    "build": "tsc",
+    "start": "node dist/index.js"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "pg": "^8.5.1"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@types/body-parser": "^1.19.0",
+    "@types/express": "^4.17.11",
+    "@types/pg": "^7.14.9",
+    "nodemon": "^2.0.7",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
+  }
 }
